refacto(api): move Prescriber read-model to prescriber-management context
