+++ api/src/shared/prescriber-management/domain/read-models/Prescriber.js
